+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add more items to the list)
 * New: Add inline documentation for hooks
 * New: Add GitHub link to readme.txt
+* Change: Add `$adjacent` argument to `{$adjacent}_or_loop_post_link`, `c2c_{$adjacent}_or_loop_post_link_get`, and `c2c_{$adjacent}_or_loop_post_link_output` filters, to sync with WP core
 * Change: Initialize plugin on 'plugins_loaded' action instead of on load
 * Change: Remove `load_textdomain()` and just load the textdomain within `init()`
 * Unit tests: